requirements.txt
langchain
langchain-google-genai
chromadb
pandas
langchain-community
langchain-classic
langchain-text-splitters
pydantic
python-dotenv
tqdm
langchain-chroma
streamlit
pypdf